feat: ffmpeg 오류 수정

+++ lib/hooks/useFFmpeg.ts
@@ -49,7 +49,7 @@ export function useFFmpeg() {
     ])
 
     const data = await ffmpeg.readFile('output.mp4')
-    return new Blob([data], { type: 'video/mp4' })
+    return new Blob([data as Uint8Array], { type: 'video/mp4' })
   }
 
   return { composeVideo, progress, loaded }
